+++ src/rs/segment.rs
@@ -105,30 +105,38 @@ impl Segment {
         None
     }
 
-    pub fn find_symbol_by_vram(&self, address: u64) -> Option<found_symbol_info::FoundSymbolInfo> {
+    pub fn find_symbol_by_vram(&self, address: u64) -> (Option<found_symbol_info::FoundSymbolInfo>, Vec<&file::File>) {
+        let mut possible_files = Vec::new();
         for file in &self.files_list {
             if let Some((sym, offset)) = file.find_symbol_by_vram(address) {
-                return Some(found_symbol_info::FoundSymbolInfo::new(
+                return (Some(found_symbol_info::FoundSymbolInfo::new(
                     file.clone(),
                     sym.clone(),
                     offset,
-                ));
+                )), Vec::new());
+            }
+            if address >= file.vram && address < file.vram + file.size {
+                possible_files.push(file);
             }
         }
-        None
+        (None, possible_files)
     }
 
-    pub fn find_symbol_by_vrom(&self, address: u64) -> Option<found_symbol_info::FoundSymbolInfo> {
+    pub fn find_symbol_by_vrom(&self, address: u64) -> (Option<found_symbol_info::FoundSymbolInfo>, Vec<&file::File>) {
+        let mut possible_files = Vec::new();
         for file in &self.files_list {
             if let Some((sym, offset)) = file.find_symbol_by_vrom(address) {
-                return Some(found_symbol_info::FoundSymbolInfo::new(
+                return (Some(found_symbol_info::FoundSymbolInfo::new(
                     file.clone(),
                     sym.clone(),
                     offset,
-                ));
+                )), Vec::new());
+            }
+            if address >= file.vram && address < file.vram + file.size {
+                possible_files.push(file);
             }
         }
-        None
+        (None, possible_files)
     }
 
     pub fn mix_folders(&self) -> Self {
@@ -419,12 +427,14 @@ pub(crate) mod python_bindings {
             self.find_symbol_by_vram_or_vrom(address)
         }
 
-        fn findSymbolByVram(&self, address: u64) -> Option<found_symbol_info::FoundSymbolInfo> {
-            self.find_symbol_by_vram(address)
+        fn findSymbolByVram(&self, address: u64) -> (Option<found_symbol_info::FoundSymbolInfo>, Vec<file::File>) {
+            let (info, possible_files) = self.find_symbol_by_vram(address);
+            (info, possible_files.into_iter().cloned().collect())
         }
 
-        fn findSymbolByVrom(&self, address: u64) -> Option<found_symbol_info::FoundSymbolInfo> {
-            self.find_symbol_by_vrom(address)
+        fn findSymbolByVrom(&self, address: u64) -> (Option<found_symbol_info::FoundSymbolInfo>, Vec<file::File>) {
+            let (info, possible_files) = self.find_symbol_by_vrom(address);
+            (info, possible_files.into_iter().cloned().collect())
         }
 
         fn mixFolders(&self) -> Self {